LayoutTests/imported/w3c/web-platform-tests/html/user-activation/detached-iframe-expected.txt


PASS navigator.userActivation retains state even if global is removed
